update and simplify files.
I'm much more proficient at using built in emacs commands and files
some of these have become less necessary. Clear them out. for local
stuff just load it like a normal person and cdb is so unpleasant the
gud add on wasn't helping much. A lot of this is made simpler by
preferring gcc and gdb when possible anyway. cmake if not and i'm fine
using the vss debugger for now. Learning cdb is much different than
lldb and gdb on first glance. Maybe I'll come back to it.