add cdb gud, always prefer debugging.
Unfortunately I have to use windows due to a work constraint. GDB does
not support pdbs and will not as long as their format is not open. This
means I have to interface with cdb. Using emacs to do so until maybe one
day pdbs are supported by cdb is my current decision